Turned the loader module into a custom type module, added basic unit tests
